backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
minio==7.2.0
aiofiles==23.2.1
python-multipart==0.0.6
passlib[bcrypt]
python-jose[cryptography]
